pyproject.toml
[project]
name = "ducklake-export"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosql>=14.1",
    "anodb>=14.0",
    "fire>=0.7.1",
    "ipython>=9.8.0",
    "jinja2>=3.1.6",
    "psycopg[binary]>=3.3.2",
    "pyathena>=3.22.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = []
